+++ history.md
@@ -7,6 +7,11 @@
 * 26
 * 27
 
+##目测比较简单,但是提交了多次
+
+* [80. Remove Duplicates from Sorted Array II]https://leetcode.com/problems/remove-duplicates-from-sorted-array-ii/
+* 
+
 ##比较困难,但是自己能够想到,之后可以再做几遍
 
 ##非常困难,不是自己解决的问题,需要再刷